LPS-94205 Merge PortalImplTest and PortalImplUnitTest since both are unit test now
