+++ portal-impl/test/unit/com/liferay/portal/util/PortalImplTest.java
@@ -14,13 +14,17 @@
 
 package com.liferay.portal.util;
 
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.security.auth.AlwaysAllowDoAsUser;
 import com.liferay.portal.kernel.servlet.PersistentHttpServletRequestWrapper;
+import com.liferay.portal.kernel.test.ReflectionTestUtil;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.upload.UploadServletRequest;
 import com.liferay.portal.kernel.util.FileUtil;
+import com.liferay.portal.kernel.util.HttpUtil;
 import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.kernel.util.PropsUtil;
+import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.upload.LiferayServletRequest;
 import com.liferay.portal.upload.UploadServletRequestImpl;
 import com.liferay.portal.util.test.PortletContainerTestUtil;
@@ -36,6 +40,7 @@ import java.util.Collections;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletRequestWrapper;
+import javax.servlet.http.HttpSession;
 
 import org.junit.Assert;
 import org.junit.BeforeClass;
@@ -64,6 +69,204 @@ public class PortalImplTest {
 		portalUtil.setPortal(new PortalImpl());
 
 		RegistryUtil.setRegistry(new BasicRegistryImpl());
+	}
+
+	@Test
+	public void testGetForwardedHost() {
+		MockHttpServletRequest mockHttpServletRequest =
+			new MockHttpServletRequest();
+
+		mockHttpServletRequest.setServerName("serverName");
+
+		Assert.assertEquals(
+			"serverName", _portalImpl.getForwardedHost(mockHttpServletRequest));
+	}
+
+	@Test
+	public void testGetForwardedHostWithCustomXForwardedHostEnabled()
+		throws Exception {
+
+		boolean webServerForwardedHostEnabled =
+			PropsValues.WEB_SERVER_FORWARDED_HOST_ENABLED;
+		String webServerForwardedHostHeader =
+			PropsValues.WEB_SERVER_FORWARDED_HOST_HEADER;
+
+		try {
+			setPropsValuesValue("WEB_SERVER_FORWARDED_HOST_ENABLED", true);
+			setPropsValuesValue(
+				"WEB_SERVER_FORWARDED_HOST_HEADER", "X-Forwarded-Custom-Host");
+
+			MockHttpServletRequest mockHttpServletRequest =
+				new MockHttpServletRequest();
+
+			mockHttpServletRequest.addHeader(
+				"X-Forwarded-Custom-Host", "forwardedServer");
+			mockHttpServletRequest.setServerName("serverName");
+
+			Assert.assertEquals(
+				"forwardedServer",
+				_portalImpl.getForwardedHost(mockHttpServletRequest));
+		}
+		finally {
+			setPropsValuesValue(
+				"WEB_SERVER_FORWARDED_HOST_ENABLED",
+				webServerForwardedHostEnabled);
+			setPropsValuesValue(
+				"WEB_SERVER_FORWARDED_HOST_HEADER",
+				webServerForwardedHostHeader);
+		}
+	}
+
+	@Test
+	public void testGetForwardedHostWithXForwardedHostDisabled()
+		throws Exception {
+
+		boolean webServerForwardedHostEnabled =
+			PropsValues.WEB_SERVER_FORWARDED_HOST_ENABLED;
+
+		try {
+			setPropsValuesValue("WEB_SERVER_FORWARDED_HOST_ENABLED", false);
+
+			MockHttpServletRequest mockHttpServletRequest =
+				new MockHttpServletRequest();
+
+			mockHttpServletRequest.addHeader(
+				"X-Forwarded-Host", "forwardedServer");
+			mockHttpServletRequest.setServerName("serverName");
+
+			Assert.assertEquals(
+				"serverName",
+				_portalImpl.getForwardedHost(mockHttpServletRequest));
+		}
+		finally {
+			setPropsValuesValue(
+				"WEB_SERVER_FORWARDED_HOST_ENABLED",
+				webServerForwardedHostEnabled);
+		}
+	}
+
+	@Test
+	public void testGetForwardedHostWithXForwardedHostEnabled()
+		throws Exception {
+
+		boolean webServerForwardedHostEnabled =
+			PropsValues.WEB_SERVER_FORWARDED_HOST_ENABLED;
+
+		try {
+			setPropsValuesValue("WEB_SERVER_FORWARDED_HOST_ENABLED", true);
+
+			MockHttpServletRequest mockHttpServletRequest =
+				new MockHttpServletRequest();
+
+			mockHttpServletRequest.addHeader(
+				"X-Forwarded-Host", "forwardedServer");
+			mockHttpServletRequest.setServerName("serverName");
+
+			Assert.assertEquals(
+				"forwardedServer",
+				_portalImpl.getForwardedHost(mockHttpServletRequest));
+		}
+		finally {
+			setPropsValuesValue(
+				"WEB_SERVER_FORWARDED_HOST_ENABLED",
+				webServerForwardedHostEnabled);
+		}
+	}
+
+	@Test
+	public void testGetForwardedPort() {
+		MockHttpServletRequest mockHttpServletRequest =
+			new MockHttpServletRequest();
+
+		mockHttpServletRequest.setServerPort(8080);
+
+		Assert.assertEquals(
+			8080, _portalImpl.getForwardedPort(mockHttpServletRequest));
+	}
+
+	@Test
+	public void testGetForwardedPortWithCustomXForwardedPort()
+		throws Exception {
+
+		boolean webServerForwardedPortEnabled =
+			PropsValues.WEB_SERVER_FORWARDED_PORT_ENABLED;
+		String webServerForwardedPortHeader =
+			PropsValues.WEB_SERVER_FORWARDED_PORT_HEADER;
+
+		try {
+			setPropsValuesValue("WEB_SERVER_FORWARDED_PORT_ENABLED", false);
+			setPropsValuesValue(
+				"WEB_SERVER_FORWARDED_PORT_HEADER", "X-Forwarded-Custom-Port");
+
+			MockHttpServletRequest mockHttpServletRequest =
+				new MockHttpServletRequest();
+
+			mockHttpServletRequest.addHeader("X-Forwarded-Custom-Port", 8081);
+			mockHttpServletRequest.setServerPort(8080);
+
+			Assert.assertEquals(
+				8080, _portalImpl.getForwardedPort(mockHttpServletRequest));
+		}
+		finally {
+			setPropsValuesValue(
+				"WEB_SERVER_FORWARDED_PORT_ENABLED",
+				webServerForwardedPortEnabled);
+			setPropsValuesValue(
+				"WEB_SERVER_FORWARDED_PORT_HEADER",
+				webServerForwardedPortHeader);
+		}
+	}
+
+	@Test
+	public void testGetForwardedPortWithXForwardedPortDisabled()
+		throws Exception {
+
+		boolean webServerForwardedHostEnabled =
+			PropsValues.WEB_SERVER_FORWARDED_PORT_ENABLED;
+
+		try {
+			setPropsValuesValue("WEB_SERVER_FORWARDED_PORT_ENABLED", false);
+
+			MockHttpServletRequest mockHttpServletRequest =
+				new MockHttpServletRequest();
+
+			mockHttpServletRequest.addHeader("X-Forwarded-Port", 8081);
+			mockHttpServletRequest.setServerPort(8080);
+
+			Assert.assertEquals(
+				8080, _portalImpl.getForwardedPort(mockHttpServletRequest));
+		}
+		finally {
+			setPropsValuesValue(
+				"WEB_SERVER_FORWARDED_PORT_ENABLED",
+				webServerForwardedHostEnabled);
+		}
+	}
+
+	@Test
+	public void testGetForwardedPortWithXForwardedPortEnabled()
+		throws Exception {
+
+		boolean webServerForwardedPortEnabled =
+			PropsValues.WEB_SERVER_FORWARDED_PORT_ENABLED;
+
+		try {
+			setPropsValuesValue("WEB_SERVER_FORWARDED_PORT_ENABLED", true);
+
+			MockHttpServletRequest mockHttpServletRequest =
+				new MockHttpServletRequest();
+
+			mockHttpServletRequest.addHeader("X-Forwarded-Port", "8081");
+			mockHttpServletRequest.setServerPort(8080);
+
+			Assert.assertEquals(
+				8081, _portalImpl.getForwardedPort(mockHttpServletRequest));
+		}
+		finally {
+			setPropsValuesValue(
+				"WEB_SERVER_FORWARDED_PORT_ENABLED",
+				webServerForwardedPortEnabled);
+		}
 	}
 
 	@Test
@@ -194,6 +397,209 @@ public class PortalImplTest {
 	}
 
 	@Test
+	public void testIsSecureWithHttpsInitialFalse() throws Exception {
+		boolean companySecurityAuthRequiresHttps =
+			PropsValues.COMPANY_SECURITY_AUTH_REQUIRES_HTTPS;
+		boolean sessionEnablePhishingProtection =
+			PropsValues.SESSION_ENABLE_PHISHING_PROTECTION;
+
+		try {
+			setPropsValuesValue("COMPANY_SECURITY_AUTH_REQUIRES_HTTPS", true);
+			setPropsValuesValue("SESSION_ENABLE_PHISHING_PROTECTION", false);
+
+			MockHttpServletRequest mockHttpServletRequest =
+				new MockHttpServletRequest();
+
+			mockHttpServletRequest.setSecure(true);
+
+			HttpSession httpSession = mockHttpServletRequest.getSession();
+
+			httpSession.setAttribute(WebKeys.HTTPS_INITIAL, Boolean.FALSE);
+
+			Assert.assertFalse(_portalImpl.isSecure(mockHttpServletRequest));
+		}
+		finally {
+			setPropsValuesValue(
+				"COMPANY_SECURITY_AUTH_REQUIRES_HTTPS",
+				companySecurityAuthRequiresHttps);
+			setPropsValuesValue(
+				"SESSION_ENABLE_PHISHING_PROTECTION",
+				sessionEnablePhishingProtection);
+		}
+	}
+
+	@Test
+	public void testIsSecureWithHttpsInitialFalseXForwardedHttps()
+		throws Exception {
+
+		boolean companySecurityAuthRequiresHttps =
+			PropsValues.COMPANY_SECURITY_AUTH_REQUIRES_HTTPS;
+		boolean sessionEnablePhishingProtection =
+			PropsValues.SESSION_ENABLE_PHISHING_PROTECTION;
+		boolean webServerForwardedProtocolEnabled =
+			PropsValues.WEB_SERVER_FORWARDED_PROTOCOL_ENABLED;
+
+		try {
+			setPropsValuesValue("COMPANY_SECURITY_AUTH_REQUIRES_HTTPS", false);
+			setPropsValuesValue("SESSION_ENABLE_PHISHING_PROTECTION", true);
+			setPropsValuesValue("WEB_SERVER_FORWARDED_PROTOCOL_ENABLED", true);
+
+			MockHttpServletRequest mockHttpServletRequest =
+				new MockHttpServletRequest();
+
+			mockHttpServletRequest.addHeader("X-Forwarded-Proto", "https");
+			mockHttpServletRequest.setSecure(false);
+
+			HttpSession httpSession = mockHttpServletRequest.getSession();
+
+			httpSession.setAttribute(WebKeys.HTTPS_INITIAL, Boolean.FALSE);
+
+			Assert.assertTrue(_portalImpl.isSecure(mockHttpServletRequest));
+		}
+		finally {
+			setPropsValuesValue(
+				"COMPANY_SECURITY_AUTH_REQUIRES_HTTPS",
+				companySecurityAuthRequiresHttps);
+			setPropsValuesValue(
+				"SESSION_ENABLE_PHISHING_PROTECTION",
+				sessionEnablePhishingProtection);
+			setPropsValuesValue(
+				"WEB_SERVER_FORWARDED_PROTOCOL_ENABLED",
+				webServerForwardedProtocolEnabled);
+		}
+	}
+
+	@Test
+	public void testIsSecureWithHttpsInitialTrue() throws Exception {
+		boolean companySecurityAuthRequiresHttps =
+			PropsValues.COMPANY_SECURITY_AUTH_REQUIRES_HTTPS;
+		boolean sessionEnablePhishingProtection =
+			PropsValues.SESSION_ENABLE_PHISHING_PROTECTION;
+
+		try {
+			setPropsValuesValue("COMPANY_SECURITY_AUTH_REQUIRES_HTTPS", true);
+			setPropsValuesValue("SESSION_ENABLE_PHISHING_PROTECTION", false);
+
+			MockHttpServletRequest mockHttpServletRequest =
+				new MockHttpServletRequest();
+
+			mockHttpServletRequest.setSecure(true);
+
+			HttpSession httpSession = mockHttpServletRequest.getSession();
+
+			httpSession.setAttribute(WebKeys.HTTPS_INITIAL, Boolean.TRUE);
+
+			Assert.assertTrue(_portalImpl.isSecure(mockHttpServletRequest));
+		}
+		finally {
+			setPropsValuesValue(
+				"COMPANY_SECURITY_AUTH_REQUIRES_HTTPS",
+				companySecurityAuthRequiresHttps);
+			setPropsValuesValue(
+				"SESSION_ENABLE_PHISHING_PROTECTION",
+				sessionEnablePhishingProtection);
+		}
+	}
+
+	@Test
+	public void testIsSecureWithHttpsInitialTrueCustomXForwardedHttps()
+		throws Exception {
+
+		boolean companySecurityAuthRequiresHttps =
+			PropsValues.COMPANY_SECURITY_AUTH_REQUIRES_HTTPS;
+		boolean sessionEnablePhishingProtection =
+			PropsValues.SESSION_ENABLE_PHISHING_PROTECTION;
+		boolean webServerForwardedProtocolEnabled =
+			PropsValues.WEB_SERVER_FORWARDED_PROTOCOL_ENABLED;
+
+		String webServerForwardedProtocolEnabledHeader =
+			PropsValues.WEB_SERVER_FORWARDED_PROTOCOL_HEADER;
+
+		try {
+			setPropsValuesValue("COMPANY_SECURITY_AUTH_REQUIRES_HTTPS", true);
+			setPropsValuesValue("SESSION_ENABLE_PHISHING_PROTECTION", false);
+			setPropsValuesValue("WEB_SERVER_FORWARDED_PROTOCOL_ENABLED", true);
+			setPropsValuesValue(
+				"WEB_SERVER_FORWARDED_PROTOCOL_HEADER",
+				"X-Forwarded-Custom-Proto");
+
+			MockHttpServletRequest mockHttpServletRequest =
+				new MockHttpServletRequest();
+
+			mockHttpServletRequest.addHeader(
+				"X-Forwarded-Custom-Proto", "https");
+			mockHttpServletRequest.setSecure(false);
+
+			Assert.assertTrue(_portalImpl.isSecure(mockHttpServletRequest));
+		}
+		finally {
+			setPropsValuesValue(
+				"COMPANY_SECURITY_AUTH_REQUIRES_HTTPS",
+				companySecurityAuthRequiresHttps);
+			setPropsValuesValue(
+				"SESSION_ENABLE_PHISHING_PROTECTION",
+				sessionEnablePhishingProtection);
+			setPropsValuesValue(
+				"WEB_SERVER_FORWARDED_PROTOCOL_ENABLED",
+				webServerForwardedProtocolEnabled);
+			setPropsValuesValue(
+				"WEB_SERVER_FORWARDED_PROTOCOL_HEADER",
+				webServerForwardedProtocolEnabledHeader);
+		}
+	}
+
+	@Test
+	public void testIsSecureWithHttpsInitialTrueXForwardedHttps()
+		throws Exception {
+
+		boolean companySecurityAuthRequiresHttps =
+			PropsValues.COMPANY_SECURITY_AUTH_REQUIRES_HTTPS;
+		boolean sessionEnablePhishingProtection =
+			PropsValues.SESSION_ENABLE_PHISHING_PROTECTION;
+		boolean webServerForwardedProtocolEnabled =
+			PropsValues.WEB_SERVER_FORWARDED_PROTOCOL_ENABLED;
+
+		try {
+			setPropsValuesValue("COMPANY_SECURITY_AUTH_REQUIRES_HTTPS", true);
+			setPropsValuesValue("SESSION_ENABLE_PHISHING_PROTECTION", false);
+			setPropsValuesValue("WEB_SERVER_FORWARDED_PROTOCOL_ENABLED", true);
+
+			MockHttpServletRequest mockHttpServletRequest =
+				new MockHttpServletRequest();
+
+			mockHttpServletRequest.addHeader("X-Forwarded-Proto", "https");
+			mockHttpServletRequest.setSecure(false);
+
+			HttpSession httpSession = mockHttpServletRequest.getSession();
+
+			httpSession.setAttribute(WebKeys.HTTPS_INITIAL, Boolean.TRUE);
+
+			Assert.assertTrue(_portalImpl.isSecure(mockHttpServletRequest));
+		}
+		finally {
+			setPropsValuesValue(
+				"COMPANY_SECURITY_AUTH_REQUIRES_HTTPS",
+				companySecurityAuthRequiresHttps);
+			setPropsValuesValue(
+				"SESSION_ENABLE_PHISHING_PROTECTION",
+				sessionEnablePhishingProtection);
+			setPropsValuesValue(
+				"WEB_SERVER_FORWARDED_PROTOCOL_ENABLED",
+				webServerForwardedProtocolEnabled);
+		}
+	}
+
+	@Test
+	public void testIsSecureWithSecureRequest() {
+		MockHttpServletRequest mockHttpServletRequest =
+			new MockHttpServletRequest();
+
+		mockHttpServletRequest.setSecure(true);
+
+		Assert.assertTrue(_portalImpl.isSecure(mockHttpServletRequest));
+	}
+
+	@Test
 	public void testIsValidResourceId() {
 		Assert.assertTrue(_portalImpl.isValidResourceId("/view.jsp"));
 		Assert.assertFalse(
@@ -210,6 +616,38 @@ public class PortalImplTest {
 		Assert.assertFalse(_portalImpl.isValidResourceId("\\WEB-INF\\web.xml"));
 	}
 
+	@Test
+	public void testUpdateRedirectRemoveLayoutURL() {
+		HttpUtil httpUtil = new HttpUtil();
+
+		httpUtil.setHttp(
+			new HttpImpl() {
+
+				@Override
+				public String getParameter(
+					String url, String name, boolean escaped) {
+
+					return StringPool.BLANK;
+				}
+
+				@Override
+				public String getPath(String url) {
+					return url;
+				}
+
+				@Override
+				public String getQueryString(String url) {
+					return StringPool.BLANK;
+				}
+
+			});
+
+		Assert.assertEquals(
+			"/web/group",
+			_portalImpl.updateRedirect(
+				"/web/group/layout", "/group/layout", "/group"));
+	}
+
 	protected HttpServletRequest getWrappedRequest(
 		HttpServletRequest requestRequest) {
 
@@ -217,6 +655,12 @@ public class PortalImplTest {
 			(HttpServletRequestWrapper)requestRequest;
 
 		return (HttpServletRequest)requestWrapper.getRequest();
+	}
+
+	protected void setPropsValuesValue(String fieldName, Object value)
+		throws Exception {
+
+		ReflectionTestUtil.setFieldValue(PropsValues.class, fieldName, value);
 	}
 
 	private void _testGetUserId(HttpServletRequest httpServletRequest) {
